cleaned up GUI module, some fixes
