+++ tmtccmd/config/args.py
@@ -295,10 +295,9 @@ def args_to_params(
     else:
         params.app_params.use_gui = pargs.gui
     if pargs.com_if is None or pargs.com_if == CoreComInterfaces.UNSPECIFIED.value:
-        if use_prompts:
-            params.com_if_id = determine_com_if(
-                hook_obj.get_com_if_dict(), hook_obj.json_cfg_path
-            )
+        params.com_if_id = determine_com_if(
+            hook_obj.get_com_if_dict(), hook_obj.json_cfg_path, use_prompts
+        )
     else:
         # TODO: Check whether COM IF is valid?
         params.com_if_id = pargs.com_if
